misc conversion warnings
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@1274 1f5c12ca-751b-0410-a591-d2e778427230

+++ branches/hieu-async/moses/src/LexicalReordering.h
@@ -72,8 +72,8 @@ private:
 	int m_orientation; /**< msd or monotone */
 	std::vector<int> m_direction;   /**< contains forward, backward, or both (bidirectional) */
 	int m_condition;   /**< fe or f */
-	int m_numScores;   /**< 1, 2, 3, or 6 */
-	int m_numOrientationTypes; /**< 2(mono) or 3(msd) */
+	size_t m_numScores;   /**< 1, 2, 3, or 6 */
+	size_t m_numOrientationTypes; /**< 2(mono) or 3(msd) */
 	std::string m_filePath; /**< probability table location */
 	vector<FactorType> m_sourceFactors; /**< source factors to condition on */
 	vector<FactorType> m_targetFactors; /**< target factors to condition on */
